split planning logic from the runner
