+++ internal/planner/discovered_tests.go
@@ -1,0 +1,73 @@
+package planner
+
+import (
+	"log/slog"
+
+	"github.com/DataDog/ddtest/internal/testoptimization"
+)
+
+func (tp *TestPlanner) processDiscoveredTests(
+	discoveredTests []testoptimization.Test,
+	skippableTests map[string]bool,
+	subdirPrefix string,
+) {
+	discoveredTestsCount := len(discoveredTests)
+	if discoveredTestsCount == 0 {
+		slog.Info("Full test discovery returned no tests")
+		return
+	}
+
+	slog.Info("Using full test discovery results")
+	skippableTestsCount := 0
+	for _, test := range discoveredTests {
+		normalizedSourceFile := stripCwdSubdirPrefix(test.SuiteSourceFile, subdirPrefix)
+		if normalizedSourceFile != "" {
+			tp.testFiles[normalizedSourceFile] = struct{}{}
+		}
+
+		if !skippableTests[test.FQN()] {
+			slog.Debug("Test is not skipped", "test", test.FQN(), "sourceFile", test.SuiteSourceFile)
+			recordRunnableTest(tp.suiteAggregates, test, normalizedSourceFile)
+		} else {
+			recordSkippedTest(tp.suiteAggregates, test, normalizedSourceFile)
+			skippableTestsCount++
+		}
+	}
+
+	slog.Info("Processed the discovered tests", "skippableTestsCount", skippableTestsCount, "discoveredTestsCount", discoveredTestsCount)
+}
+
+func (tp *TestPlanner) processDiscoveredTestFiles(discoveredTestFiles []string) {
+	for _, testFile := range discoveredTestFiles {
+		if testFile != "" {
+			tp.testFiles[testFile] = struct{}{}
+		}
+	}
+}
+
+func recordRunnableTest(suiteAggregates map[testSuiteKey]testSuiteAggregate, test testoptimization.Test, sourceFile string) {
+	aggregate := suiteAggregateForTest(suiteAggregates, test, sourceFile)
+	aggregate.NumTests++
+	suiteAggregates[testSuiteKey{Module: test.Module, Suite: test.Suite}] = aggregate
+}
+
+func recordSkippedTest(suiteAggregates map[testSuiteKey]testSuiteAggregate, test testoptimization.Test, sourceFile string) {
+	aggregate := suiteAggregateForTest(suiteAggregates, test, sourceFile)
+	aggregate.NumTests++
+	aggregate.NumTestsSkipped++
+	suiteAggregates[testSuiteKey{Module: test.Module, Suite: test.Suite}] = aggregate
+}
+
+func suiteAggregateForTest(suiteAggregates map[testSuiteKey]testSuiteAggregate, test testoptimization.Test, sourceFile string) testSuiteAggregate {
+	key := testSuiteKey{
+		Module: test.Module,
+		Suite:  test.Suite,
+	}
+	aggregate := suiteAggregates[key]
+	if aggregate.SourceFile == "" {
+		aggregate.Module = test.Module
+		aggregate.Suite = test.Suite
+		aggregate.SourceFile = sourceFile
+	}
+	return aggregate
+}
