+++ internal/planner/parallelism_test.go
@@ -1,4 +1,4 @@
-package runner
+package planner
 
 import (
 	"encoding/json"
