+++ internal/runner/ci_node_executor.go
@@ -4,13 +4,15 @@ import (
 	"fmt"
 	"log/slog"
 	"os"
+	"path/filepath"
 
+	"github.com/DataDog/ddtest/internal/constants"
 	"golang.org/x/sync/errgroup"
 )
 
 // runCINode executes tests for a specific CI node (one split, not the whole tests set).
 // It further splits the node's tests among local workers based on ci_node_workers setting.
-func (e testExecutor) runCINode(ciNode int, ciNodeWorkers int, testFileWeights map[string]int) runExecutionResult {
+func (e testExecutor) runCINode(ciNode int, ciNodeWorkers int) runExecutionResult {
 	report := newCINodeExecutionReport(ciNode, ciNodeWorkers)
 	testFiles, err := loadCINodeTestFiles(ciNode)
 	if err != nil {
@@ -21,7 +23,7 @@ func (e testExecutor) runCINode(ciNode int, ciNodeWorkers int, testFileWeights m
 	if report.LocalWorkers <= 1 {
 		err = e.runCINodeSingleWorker(ciNode, testFiles)
 	} else {
-		err = e.runCINodeWorkers(ciNode, report.LocalWorkers, testFiles, testFileWeights)
+		err = e.runCINodeWorkers(ciNode, report.LocalWorkers, testFiles)
 	}
 	if err != nil {
 		return report.failure(err)
@@ -42,7 +44,7 @@ func newCINodeExecutionReport(ciNode int, ciNodeWorkers int) runExecutionReport 
 }
 
 func loadCINodeTestFiles(ciNode int) ([]string, error) {
-	runnerFilePath := runnerSplitPath(ciNode)
+	runnerFilePath := filepath.Join(constants.TestsSplitDir, fmt.Sprintf("runner-%d", ciNode))
 	testFiles, err := loadTestBatch(runnerFilePath)
 	if os.IsNotExist(err) {
 		return nil, fmt.Errorf("runner file for ci-node %d does not exist: %s", ciNode, runnerFilePath)
@@ -62,7 +64,7 @@ func (e testExecutor) runCINodeSingleWorker(ciNode int, testFiles []string) erro
 	return e.runBatch(testFiles, ciNode, 0)
 }
 
-func (e testExecutor) runCINodeWorkers(ciNode int, ciNodeWorkers int, testFiles []string, testFileWeights map[string]int) error {
+func (e testExecutor) runCINodeWorkers(ciNode int, ciNodeWorkers int, testFiles []string) error {
 	if len(testFiles) == 0 {
 		slog.Info("No tests to run for CI node", "ciNode", ciNode)
 		return nil
@@ -71,10 +73,7 @@ func (e testExecutor) runCINodeWorkers(ciNode int, ciNodeWorkers int, testFiles 
 	slog.Info("Running tests for CI node in parallel mode",
 		"ciNode", ciNode, "ciNodeWorkers", ciNodeWorkers, "testFilesCount", len(testFiles))
 
-	if testFileWeights == nil {
-		testFileWeights = map[string]int{}
-	}
-	groups := subsplitTestsBetweenWorkers(testFiles, ciNodeWorkers, testFileWeights)
+	groups := e.subsplitTestsBetweenWorkers(testFiles, ciNodeWorkers)
 	return e.runCINodeWorkerGroups(ciNode, groups)
 }
 
@@ -101,21 +100,6 @@ func (e testExecutor) runCINodeWorkerGroups(ciNode int, groups [][]string) error
 	return nil
 }
 
-// subsplitTestsBetweenWorkers splits a CI node's test files among local workers
-// using the same weighted distribution algorithm used for CI node splits.
-func subsplitTestsBetweenWorkers(testFiles []string, n int, testFileWeights map[string]int) [][]string {
-	if n <= 0 {
-		n = 1
-	}
-
-	nodeTestFiles := make(map[string]int, len(testFiles))
-	for _, testFile := range testFiles {
-		if cachedWeight, ok := testFileWeights[testFile]; ok && cachedWeight > 0 {
-			nodeTestFiles[testFile] = cachedWeight
-		} else {
-			nodeTestFiles[testFile] = defaultTestFileWeight
-		}
-	}
-
-	return DistributeTestFiles(nodeTestFiles, n)
+func (e testExecutor) subsplitTestsBetweenWorkers(testFiles []string, n int) [][]string {
+	return e.planner.DistributeTestFiles(testFiles, n)
 }
