+++ internal/runner/runner.go
@@ -2,159 +2,56 @@ package runner
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"io"
 	"log/slog"
 	"os"
-	"slices"
 	"strings"
 	"time"
 
 	ciUtils "github.com/DataDog/ddtest/civisibility/utils"
-	"github.com/DataDog/ddtest/internal/ciprovider"
 	"github.com/DataDog/ddtest/internal/constants"
+	"github.com/DataDog/ddtest/internal/planner"
 	"github.com/DataDog/ddtest/internal/platform"
+	"github.com/DataDog/ddtest/internal/runmetadata"
 	"github.com/DataDog/ddtest/internal/settings"
-	"github.com/DataDog/ddtest/internal/testoptimization"
 )
 
 type Runner interface {
-	Plan(ctx context.Context) error
 	Run(ctx context.Context) error
 }
 
+type Planner interface {
+	Plan(ctx context.Context) error
+	LoadPlan() (planner.PlanInfo, error)
+	DistributeTestFiles(testFiles []string, parallelRunners int) [][]string
+}
+
 type TestRunner struct {
-	testFiles               map[string]struct{}
-	suiteAggregates         map[testSuiteKey]testSuiteAggregate
-	suitesBySourceFile      map[string][]testSuiteKey
-	testSuiteDurations      map[string]map[string]testoptimization.TestSuiteDurationInfo
-	testFileWeights         map[string]int
-	testFileDurationSources map[string]testFileDurationSource
-	skippablePercentage     float64
-	planReport              planReport
-	runInfoReport           runInfoReport
-	platformDetector        platform.PlatformDetector
-	optimizationClient      testoptimization.TestOptimizationClient
-	durationsClient         testoptimization.TestSuiteDurationsClient
-	ciProviderDetector      ciprovider.CIProviderDetector
-	reportWriter            io.Writer
+	platformDetector platform.PlatformDetector
+	planner          Planner
+	reportWriter     io.Writer
 }
 
 func New() *TestRunner {
-	runner := newTestRunnerWithDefaults()
-	runner.platformDetector = platform.NewPlatformDetector()
-	runner.optimizationClient = testoptimization.NewDatadogClient()
-	runner.durationsClient = testoptimization.NewDurationsClient()
-	runner.ciProviderDetector = ciprovider.NewCIProviderDetector()
-	return runner
+	return NewWithDependencies(platform.NewPlatformDetector(), planner.New())
 }
 
 func NewWithDependencies(
 	platformDetector platform.PlatformDetector,
-	optimizationClient testoptimization.TestOptimizationClient,
-	durationsClient testoptimization.TestSuiteDurationsClient,
-	ciProviderDetector ciprovider.CIProviderDetector,
+	testPlanner Planner,
 ) *TestRunner {
 	runner := newTestRunnerWithDefaults()
 	runner.platformDetector = platformDetector
-	runner.optimizationClient = optimizationClient
-	runner.durationsClient = durationsClient
-	runner.ciProviderDetector = ciProviderDetector
+	runner.planner = testPlanner
 	return runner
 }
 
 func newTestRunnerWithDefaults() *TestRunner {
 	return &TestRunner{
-		testFiles:               make(map[string]struct{}),
-		suiteAggregates:         make(map[testSuiteKey]testSuiteAggregate),
-		suitesBySourceFile:      make(map[string][]testSuiteKey),
-		testSuiteDurations:      make(map[string]map[string]testoptimization.TestSuiteDurationInfo),
-		testFileWeights:         make(map[string]int),
-		testFileDurationSources: make(map[string]testFileDurationSource),
-		skippablePercentage:     0.0,
-		reportWriter:            os.Stderr,
+		planner:      planner.New(),
+		reportWriter: os.Stderr,
 	}
-}
-
-func (tr *TestRunner) Plan(ctx context.Context) error {
-	slog.Info("Planning test execution...")
-
-	if err := tr.PrepareTestOptimization(ctx); err != nil {
-		return err
-	}
-
-	if err := writePlanFile(constants.ManifestPath, []byte(constants.ManifestVersion+"\n")); err != nil {
-		return fmt.Errorf("failed to write test optimization manifest: %w", err)
-	}
-
-	if err := tr.storeTestOptimizationPlanCache(); err != nil {
-		return fmt.Errorf("failed to store test optimization plan cache: %w", err)
-	}
-
-	testFileNames := make([]string, 0, len(tr.testFileWeights))
-	for testFile := range tr.testFileWeights {
-		testFileNames = append(testFileNames, testFile)
-	}
-	slices.Sort(testFileNames)
-
-	content := strings.Join(testFileNames, "\n")
-	if len(testFileNames) > 0 {
-		content += "\n"
-	}
-
-	if err := writePlanFile(constants.TestFilesOutputPath, []byte(content)); err != nil {
-		return fmt.Errorf("failed to write test files: %w", err)
-	}
-
-	percentageContent := fmt.Sprintf("%.2f", tr.skippablePercentage)
-	if err := writePlanFile(constants.SkippablePercentageOutputPath, []byte(percentageContent)); err != nil {
-		return fmt.Errorf("failed to write skippable percentage: %w", err)
-	}
-
-	// Calculate and write parallel runners count
-	parallelRunnerSplit := calculateParallelRunnerSplit(
-		tr.testFileWeights,
-		settings.GetMinParallelism(),
-		settings.GetMaxParallelism(),
-		settings.GetParallelRunnerOverhead(),
-	)
-	parallelRunners := parallelRunnerSplit.parallelRunners
-	runnersContent := fmt.Sprintf("%d", parallelRunners)
-	if err := writePlanFile(constants.ParallelRunnersOutputPath, []byte(runnersContent)); err != nil {
-		return fmt.Errorf("failed to write parallel runners: %w", err)
-	}
-
-	// Detect and configure CI provider if available
-	if ciProvider, err := tr.ciProviderDetector.DetectCIProvider(); err == nil {
-		slog.Info("CI provider detected, configuring with parallel runners",
-			"provider", ciProvider.Name(), "parallelRunners", parallelRunners)
-
-		if err := ciProvider.Configure(parallelRunners); err != nil {
-			slog.Warn("Failed to configure CI provider", "provider", ciProvider.Name(), "error", err)
-		}
-	} else {
-		slog.Info("No CI provider detected or CI provider is not supported, running tests without CI integration", "error", err)
-	}
-
-	// Split test files for runners
-	if err := CreateTestSplits(tr.testFileWeights, parallelRunners, constants.TestFilesOutputPath); err != nil {
-		return fmt.Errorf("failed to create test splits: %w", err)
-	}
-
-	tr.planReport.Split = parallelRunnerSplit
-	slog.Info("Test execution planning completed",
-		"parallelRunners", parallelRunners,
-		"expectedWallTime", parallelRunnerSplit.wallTimeDuration(),
-		"imbalance", parallelRunnerSplit.imbalanceDuration(),
-		"expectedTotalRuntime", parallelRunnerSplit.totalRuntimeDuration(),
-		"testFilesCount", len(tr.testFileWeights))
-
-	if settings.GetReportEnabled() {
-		printPlanReport(tr.reportWriter, tr.planReport)
-	}
-
-	return nil
 }
 
 func (tr *TestRunner) Run(ctx context.Context) error {
@@ -163,34 +60,23 @@ func (tr *TestRunner) Run(ctx context.Context) error {
 		slog.Info("Test optimization planning data not found, running planning phase...")
 
 		// Run Setup if the file doesn't exist
-		if err := tr.Plan(ctx); err != nil {
+		if err := tr.planner.Plan(ctx); err != nil {
 			return fmt.Errorf("failed to run planning phase: %w", err)
 		}
 	} else if err != nil {
 		return fmt.Errorf("failed to check parallel runners count at %s: %w", constants.ParallelRunnersOutputPath, err)
 	}
 
-	if err := tr.restoreTestOptimizationPlanCache(); err != nil {
-		if errors.Is(err, os.ErrNotExist) {
-			slog.Debug("Test optimization plan cache not found; CI-node subsplits will use default weights",
-				"file", testoptimization.TestOptimizationPlanCacheFile)
-		} else {
-			slog.Warn("Failed to restore test optimization plan cache; CI-node subsplits will use default weights",
-				"file", testoptimization.TestOptimizationPlanCacheFile, "error", err)
-		}
-	}
-
-	runnersData, err := os.ReadFile(constants.ParallelRunnersOutputPath)
+	planInfo, err := tr.planner.LoadPlan()
 	if err != nil {
-		return fmt.Errorf("failed to read parallel runners count from %s: %w", constants.ParallelRunnersOutputPath, err)
-	}
-	runnersString := strings.TrimSpace(string(runnersData))
-
-	parallelRunners := 0
-	if _, err := fmt.Sscanf(runnersString, "%d", &parallelRunners); err != nil {
-		return fmt.Errorf("failed to parse parallel runners count from %s: %w", runnersString, err)
+		slog.Error("Test optimization plan is not available", "error", err)
+		return fmt.Errorf("test optimization plan is not available: %w", err)
 	}
 
+	parallelRunners, err := readParallelRunnersCount()
+	if err != nil {
+		return err
+	}
 	slog.Info("Got parallel runners count", "parallelRunners", parallelRunners)
 
 	// Parse worker environment variables if provided in settings
@@ -209,16 +95,17 @@ func (tr *TestRunner) Run(ctx context.Context) error {
 		return fmt.Errorf("failed to detect framework: %w", err)
 	}
 	slog.Info("Framework detected", "framework", framework.Name())
-	if tr.runInfoReport.isZero() {
-		tr.runInfoReport = newRunInfoReport(ciUtils.GetCITags(), nil, detectedPlatform.Name(), framework.Name())
+	runInfo := runmetadata.New(ciUtils.GetCITags())
+	if planInfo.IsZero() {
+		planInfo = planner.NewPlanInfo(nil, detectedPlatform.Name(), framework.Name())
 	}
 
 	ciNode := settings.GetCiNode()
 	startTime := time.Now()
-	executor := newTestExecutor(ctx, framework, workerEnvMap)
+	executor := newTestExecutor(ctx, framework, workerEnvMap, tr.planner)
 	var executionResult runExecutionResult
 	if ciNode >= 0 {
-		executionResult = executor.runCINode(ciNode, settings.GetCiNodeWorkers(), tr.testFileWeights)
+		executionResult = executor.runCINode(ciNode, settings.GetCiNodeWorkers())
 	} else if parallelRunners > 1 {
 		executionResult = executor.runParallel()
 	} else {
@@ -227,11 +114,27 @@ func (tr *TestRunner) Run(ctx context.Context) error {
 
 	if settings.GetReportEnabled() {
 		printRunReport(tr.reportWriter, runReport{
-			RunInfo:   tr.runInfoReport,
+			RunInfo:   runInfo,
+			PlanInfo:  planInfo,
 			Execution: executionResult.report,
 			Duration:  time.Since(startTime),
 			Err:       executionResult.err,
 		})
 	}
 	return executionResult.err
+}
+
+func readParallelRunnersCount() (int, error) {
+	runnersData, err := os.ReadFile(constants.ParallelRunnersOutputPath)
+	if err != nil {
+		return 0, fmt.Errorf("failed to read parallel runners count from %s: %w", constants.ParallelRunnersOutputPath, err)
+	}
+	runnersString := strings.TrimSpace(string(runnersData))
+
+	parallelRunners := 0
+	if _, err := fmt.Sscanf(runnersString, "%d", &parallelRunners); err != nil {
+		return 0, fmt.Errorf("failed to parse parallel runners count from %s: %w", runnersString, err)
+	}
+
+	return parallelRunners, nil
 }
