+++ internal/runner/worker_env_test.go
@@ -57,7 +57,7 @@ func TestRunBatch_DefaultTestSessionName(t *testing.T) {
 		RunTestsCalls: []RunTestsCall{},
 	}
 
-	err := newTestExecutor(context.Background(), mockFramework, map[string]string{}).runBatch([]string{"test/file1_test.rb"}, 2, 4)
+	err := newTestExecutor(context.Background(), mockFramework, map[string]string{}, roundRobinTestPlanner{}).runBatch([]string{"test/file1_test.rb"}, 2, 4)
 	if err != nil {
 		t.Fatalf("runBatch() should not return error, got: %v", err)
 	}
@@ -81,7 +81,7 @@ func TestRunBatch_DefaultTestSessionNameUsesDDService(t *testing.T) {
 		RunTestsCalls: []RunTestsCall{},
 	}
 
-	err := newTestExecutor(context.Background(), mockFramework, map[string]string{}).runBatch([]string{"test/file1_test.rb"}, 3, 7)
+	err := newTestExecutor(context.Background(), mockFramework, map[string]string{}, roundRobinTestPlanner{}).runBatch([]string{"test/file1_test.rb"}, 3, 7)
 	if err != nil {
 		t.Fatalf("runBatch() should not return error, got: %v", err)
 	}
@@ -103,7 +103,7 @@ func TestRunBatch_UserTestSessionNameSupportsPlaceholders(t *testing.T) {
 		RunTestsCalls: []RunTestsCall{},
 	}
 
-	err := newTestExecutor(context.Background(), mockFramework, map[string]string{}).runBatch([]string{"test/file1_test.rb"}, 5, 8)
+	err := newTestExecutor(context.Background(), mockFramework, map[string]string{}, roundRobinTestPlanner{}).runBatch([]string{"test/file1_test.rb"}, 5, 8)
 	if err != nil {
 		t.Fatalf("runBatch() should not return error, got: %v", err)
 	}
@@ -128,7 +128,7 @@ func TestRunBatch_WorkerEnvTestSessionNameTakesPrecedence(t *testing.T) {
 		ciConstants.CIVisibilityTestSessionNameEnvironmentVariable: "worker-node-{{nodeIndex}}-worker-{{workerIndex}}",
 	}
 
-	err := newTestExecutor(context.Background(), mockFramework, workerEnvMap).runBatch([]string{"test/file1_test.rb"}, 9, 1)
+	err := newTestExecutor(context.Background(), mockFramework, workerEnvMap, roundRobinTestPlanner{}).runBatch([]string{"test/file1_test.rb"}, 9, 1)
 	if err != nil {
 		t.Fatalf("runBatch() should not return error, got: %v", err)
 	}
@@ -150,7 +150,7 @@ func TestRunBatch_DefaultManifestFile(t *testing.T) {
 		RunTestsCalls: []RunTestsCall{},
 	}
 
-	err := newTestExecutor(context.Background(), mockFramework, map[string]string{}).runBatch([]string{"test/file1_test.rb"}, 2, 4)
+	err := newTestExecutor(context.Background(), mockFramework, map[string]string{}, roundRobinTestPlanner{}).runBatch([]string{"test/file1_test.rb"}, 2, 4)
 	if err != nil {
 		t.Fatalf("runBatch() should not return error, got: %v", err)
 	}
@@ -181,7 +181,7 @@ func TestRunBatch_ManifestFileUsesProcessEnv(t *testing.T) {
 		RunTestsCalls: []RunTestsCall{},
 	}
 
-	err := newTestExecutor(context.Background(), mockFramework, map[string]string{}).runBatch([]string{"test/file1_test.rb"}, 2, 4)
+	err := newTestExecutor(context.Background(), mockFramework, map[string]string{}, roundRobinTestPlanner{}).runBatch([]string{"test/file1_test.rb"}, 2, 4)
 	if err != nil {
 		t.Fatalf("runBatch() should not return error, got: %v", err)
 	}
@@ -203,7 +203,7 @@ func TestRunBatch_WorkerEnvManifestFileTakesPrecedence(t *testing.T) {
 		constants.TestOptimizationManifestFileEnvVar: "/tmp/worker-manifest.txt",
 	}
 
-	err := newTestExecutor(context.Background(), mockFramework, workerEnvMap).runBatch([]string{"test/file1_test.rb"}, 2, 4)
+	err := newTestExecutor(context.Background(), mockFramework, workerEnvMap, roundRobinTestPlanner{}).runBatch([]string{"test/file1_test.rb"}, 2, 4)
 	if err != nil {
 		t.Fatalf("runBatch() should not return error, got: %v", err)
 	}
